requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6      
exceptiongroup==1.1.3
fastapi==0.101.1     
h11==0.14.0
idna==3.4
paho-mqtt==1.6.1     
pydantic==1.9.0      
pydantic-core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.2